supabase対応

+++ app/admin/posts/[id]/page.tsx
@@ -1,22 +1,27 @@
 "use client";
 
-import { useEffect, useState, type SubmitEvent } from "react";
+import { ChangeEvent, useEffect, useState, type SubmitEvent } from "react";
 import { useParams, useRouter } from "next/navigation";
+import { v4 as uuidv4 } from "uuid";
+import { supabase } from "@/app/_libs/supabase";
 import type { GetPostsIdResponse } from "@/app/api/posts/[id]/route";
 import type { UpdatePostRequestBody } from "@/app/api/admin/posts/[id]/route";
 import type { CategoriesResponse } from "@/app/api/admin/categories/route";
 import PostForm from "@/app/_components/admin/PostForm";
+import { useSupabaseSession } from "@/app/_hooks/useSupabaseSession";
 
 export default function AdminPostIdPage() {
 
   const router = useRouter();
   const { id } = useParams();
+  const { token } = useSupabaseSession();
   const [ loading, setLoading ] = useState(true);
   const [ error, setError ]     = useState<string | null>(null);
 
   const [ title, setTitle ] = useState("");
   const [ content, setContent ] = useState("");
-  const [ thumbnailUrl, setThumbnailUrl ] = useState("");
+  const [ thumbnailImageKey, setThumbnailImageKey ] = useState("");
+  const [ thumbnailImageUrl, setThumbnailImageUrl ] = useState<null | string>(null);
   const [ categories, setCategories ]     = useState<CategoriesResponse["categories"]>([]);
   const [ selectCategories, setSelectCategories ] = useState<number[]>([]);
   const [ isSubmitting, setIsSubmitting ] = useState(false);
@@ -25,8 +30,15 @@ export default function AdminPostIdPage() {
   useEffect(() => {
     const fetcher = async () => {
 
+      if (!token) return;
+
       try {
-        const res = await fetch("/api/admin/categories");
+        const res = await fetch("/api/admin/categories", {
+          headers: {
+            "Content-Type": "application/json",
+            Authorization: token,
+          },
+        });
         const { categories } = await res.json();
         setCategories(categories);
 
@@ -37,11 +49,16 @@ export default function AdminPostIdPage() {
       }
 
       try {
-        const res = await fetch(`/api/admin/posts/${id}`);
+        const res = await fetch(`/api/admin/posts/${id}`, {
+          headers: {
+            "Content-Type": "application/json",
+            Authorization: token,
+          },
+        });
         const { post } = await res.json() as GetPostsIdResponse;
         setTitle(post.title);
         setContent(post.content);
-        setThumbnailUrl(post.thumbnailUrl);
+        setThumbnailImageKey(post.thumbnailImageKey);
         setSelectCategories(
           post.postCategories.map((postCategory) => postCategory.category.id)
         );
@@ -55,7 +72,24 @@ export default function AdminPostIdPage() {
     }
 
     fetcher();
-  }, []);
+  }, [token, id]);
+
+  // thumbnailImageKeyを用いて画像のURLを取得
+  useEffect(() => {
+    if (!thumbnailImageKey) return;
+
+    const fetcher = async () => {
+      const {
+        data: { publicUrl },
+      } = await supabase.storage
+        .from('post_thumbnail')
+        .getPublicUrl(thumbnailImageKey);
+
+      setThumbnailImageUrl(publicUrl);
+    };
+
+    fetcher();
+  }, [thumbnailImageKey]);
 
   // 読み込み中表示
   if(loading) {
@@ -76,13 +110,15 @@ export default function AdminPostIdPage() {
   }
 
   // 記事更新
-  const handleSubmit = async (e: SubmitEvent<HTMLFormElement>) => { 
+  const handleSubmit = async (e: SubmitEvent<HTMLFormElement>) => {
     e.preventDefault();
 
-    const body: UpdatePostRequestBody = { 
+    if (!token) return;
+
+    const body: UpdatePostRequestBody = {
       title,
       content,
-      thumbnailUrl, 
+      thumbnailImageKey,
       categories: selectCategories.map((id) => ({ id })),
     };
 
@@ -91,7 +127,10 @@ export default function AdminPostIdPage() {
     try {
       await fetch(`/api/admin/posts/${id}`, {
         method: 'PUT',
-        headers: { 'Content-Type': 'application/json' },
+        headers: {
+          'Content-Type': 'application/json',
+          Authorization: token,
+        },
         body: JSON.stringify(body),
       });
 
@@ -106,14 +145,20 @@ export default function AdminPostIdPage() {
   }
 
   // 記事削除
-  const handleDelete = async () => { 
+  const handleDelete = async () => {
     if (!confirm('削除しますか？')) return;
+
+    if (!token) return;
 
     setIsSubmitting(true);
 
     try {
       await fetch(`/api/admin/posts/${id}`, {
         method: 'DELETE',
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: token,
+        },
       });
 
       alert("削除しました");
@@ -127,6 +172,30 @@ export default function AdminPostIdPage() {
     }
   }
 
+  // サムネイル画像設定
+  const handleImageChange = async (e: ChangeEvent<HTMLInputElement>): Promise<void> => {
+    if (!e.target.files || e.target.files.length == 0) {
+      return;
+    }
+
+    const file = e.target.files[0];
+    const filePath = `private/${uuidv4()}`;
+
+    const { data, error } = await supabase.storage
+      .from('post_thumbnail')
+      .upload(filePath, file, {
+        cacheControl: '3600',
+        upsert: false,
+      });
+
+    if (error) {
+      alert(error.message);
+      return;
+    }
+
+    setThumbnailImageKey(data.path);
+  }
+
   return (
     <div>
       <h1 className="text-2xl font-bold">記事作成</h1>
@@ -136,8 +205,6 @@ export default function AdminPostIdPage() {
         setTitle={setTitle}
         content={content}
         setContent={setContent}
-        thumbnailUrl={thumbnailUrl}
-        setThumbnailUrl={setThumbnailUrl}
         categories={categories}
         selectCategories={selectCategories}
         setSelectCategories={setSelectCategories}
@@ -145,6 +212,8 @@ export default function AdminPostIdPage() {
         submitLabel="更新"
         onSubmit={handleSubmit}
         onDelete={handleDelete}
+        handleImageChange={handleImageChange}
+        thumbnailImageUrl={thumbnailImageUrl}
       />
     </div>
   );
